long long to ascii function --> intmax_t to ascii and uintmax_t to ascii base

+++ libft/sources/ft_itoa_base.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	change_base(long long n, int base, char *str, int i)
+void	change_base(uintmax_t n, int base, char *str, int i)
 {
 	int	rest;
